First Commit - Level Editor and Grid System
GridXY is a generic class that allows to create grid in XY axis.

Grid Logic System is responsible for setting up the level from levelSO scriptable object that is made by Level Editor Scene. It allows user to set level design visually.

Grid Visual System is responsible for representing candies in the grid cell visually.

The other scripts will be explained in readme in the end of the project.